requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
ale-py==0.8.1
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
colorama==0.4.6
commentjson==0.9.0
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
etils==0.8.0
exceptiongroup==1.0.1
executing==1.1.1
Farama-Notifications==0.0.4
ffmpeg==1.4
fire==0.4.0
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.50.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
h5py==3.9.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
ipykernel==6.16.1
ipython==8.5.0
jax==0.3.23
jax-jumpy==1.0.0
jaxlib==0.3.22+cuda11.cudnn82
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.4.3
jupyter_core==4.11.2
keras==2.12.0
kiwisolver==1.4.4
lark-parser==0.7.8
libclang==16.0.0
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
numpy==1.23.4
numpy-stl==3.0.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
-e git+ssh://git@github.com/Rhoban/onshape-to-robot.git@1d1062021df97dedeecd5ee72ff051cf0f52f4b8#egg=onshape_to_robot
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==4.23.3
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pygame==2.1.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-utils==3.7.0
pytorch-lightning==1.8.0.post1
pytz==2023.3
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
six==1.16.0
stable-baselines==2.10.2
stack-data==0.5.1
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.1.0
tomli==2.0.1
torch==1.13.0
torchmetrics==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transforms3d==0.4.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.10.0
